// src/app/components/merchandising/Costing/model/costing-model.ts
import { CostAccessories } from "./costing-accessories.model";
import { CostFabric } from "./costing-fabric-model";

export class CostingMasterModel {
    id: number;
    costNumber: number;
    initialDate: Date;
    entryDate: Date;
    versionNo: number;
    factoryId: number;
    branceOfficeName: string;
    buyerId: number;
    buyerName: string;
    buyerDeptId: number;
    deptName: string;
    buyerCP: string;
    localAgency: string;
    colorType: string;
    noOfColor: number;
    sizeRangeId: number;
    sizeRange: string;
    costSize: string;
    marchandiserId: number;
    accountHolderId: number;
    styleId: number;
    styleName: string;
    item: string;
    description: string;
    styleNo: string;
    seasonId: number;
    seasonName: string;
    yearId: number;
    yearName: string;
    firstDeliveryDate: Date;
    moq: number;
    mcq: number;
    estimateQty: number;
    deliveryMode: string;
    paymentMode: string;
    tenor: number;
    priceTerm: string;
    foreignCommPc: number;
    localCommPc: number;
    specialCommPc: number;
    currencyId: number;
    currencyName:string;
    currencyConversionRate: number;
    program: string;
    packingType: string;
    specialNote: string;
    smv: number;
    ppm: number;
    efficency: number;
    cmPC: number;
    cmDz: number;
    confDate: Date;
    confRate: number;
    // New 
    countryId: number;
    countryName: string;
    brandId: number;
    brandName: string;
    shippedTypeId: number;
    shippedTypeName: string;
    fobPc: number;
    fobDz: number;
    poq: number;
    coq: number;
    buyerDesign: string;
    preFilledInformStyle: string;
    departmentNo: string;
    departmentDescription: string;
    costingFabricsList:CostFabric[];
    costAccessoriesList:CostAccessories[];
    costPackingingList:CostAccessories[];
    costLabelList:CostAccessories[];
    costProcessList:CostAccessories[];

    metarialCost: number;
    metarialCostPc: number;
    finalCMDZ: number;
    finalCMPc: number;
    finalCommission: number;
    finalCommissionPc: number;
    surPlus: number;
    surPlusPc: number;
    netFOB: number;
    netFOBPc: number;
    surPlusPercentage: number;
    finalCommissionPercentage: number;
    otherDz: number;
    otherPc: number;
    filePath: File;
    image:string;
    costOption: string;
    quotationNumber: string;
    showQuotationNumber: string;
    status: string;
    metarialCostPercentage: number;
    othersPercentage: number;
    finalCMPercentage: number;
    colorCode: string;
    employeeName: string;
    revisedNo: number;
    quotationRevised: string;
    targetFOB: number;
    stylePartId: number;
    checkedDate: Date;
    checkedBy: string;
    checkedByName: string;
    approvedBy: string;
    approvedDate: Date;
    approvedByName: string;
    updateDate: string;
    updateBy: string;
    partName: string;
    costingStatus: string;
    updateFrom: string;
    embroidery: number;
    leaser: number;
    printing: number;
    quilting: number;
    washing: number;
    cpm: number;
    factorialCostPc: number;
    factorialCostPercentage: number;
    factorialCostDz: number;
    fabAccessCount: number;
    refStyleId: number;
    styleStatus: string;
    preStatus: string;
    editInfo: string;
    preFromQuotation: string;
  }
  